requirements.txt
amqp==5.2.0
asgiref==3.7.2
autopep8==2.0.4
billiard==4.2.0
blinker==1.7.0
Brotli==1.1.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
coverage==7.3.2
Django==4.2.7
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-js-asset==2.1.0
django-redis==5.4.0
django-silk==5.0.4
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
flower==2.0.1
gevent==23.9.1
geventhttpclient==2.0.11
gprof2dot==2022.7.29
greenlet==3.0.1
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.4
locust==2.19.0
MarkupSafe==2.1.3
model-bakery==1.17.0
msgpack==1.0.7
Pillow==10.1.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
psutil==5.9.6
pycodestyle==2.11.1
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
redis==5.0.1
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
tornado==6.3.3
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
